book 조건 추가

+++ ZIBI/src/main/java/kr/spring/book/controller/BookAjaxController.java
@@ -63,48 +63,54 @@ public class BookAjaxController {
 										@RequestParam int book_num,
 										@RequestParam int apply_num,
 										@RequestParam int book_state,
-										@RequestParam String apply_gatheringDate){
+										@RequestParam String apply_gatheringDate,
+										HttpSession session){
 		
 		Map<String,String> mapJson = new HashMap<String,String>();
 		
-		//일정 중복 여부 체크
-		BookMatchingVO db_match = bookService.selectMatch(apply_gatheringDate,apply_num);
-		if(db_match!=null) {
-			mapJson.put("result", "duplicated");
+		MemberVO user = (MemberVO)session.getAttribute("user");
+		if(user==null) {
+			mapJson.put("result", "logout");
 		}else {
-			//예약 신청 정보 세팅
-			BookMatchingVO bookMatchingVO = new BookMatchingVO();
-			bookMatchingVO.setBook_num(book_num);
-			bookMatchingVO.setApply_num(apply_num);
-			bookMatchingVO.setBook_state(book_state);
-			bookMatchingVO.setApply_gatheringDate(apply_gatheringDate);
-			
-			//예약 신청 완료
-			bookService.insertMatch(bookMatchingVO);
-			
-			//이메일 내용 세팅
-			MemberVO member = memberService.selectMember(apply_num);
-			String nickname = member.getMem_nickname();
-			
-			//예약 완료 이메일 양식
-			String setFrom = "229rkawk@gmail.com";
-			String toMail = email;
-			String title = "[ZIBI] 소모임 참여 신청 안내";
-			String content = nickname + "님의 소모임 예약이 완료되었습니다.<br>";
-			content += "변경 사항이 발생할 수 있으니 ZIBI [나의 모임]에서 꼭 확인하세요!";
-			
-			try {
-				MimeMessage message = mailSender.createMimeMessage();
-				MimeMessageHelper helper = new MimeMessageHelper(message, true, "utf-8");
-				helper.setFrom(setFrom);
-				helper.setTo(toMail);
-				helper.setSubject(title);
-				helper.setText(content, true);
-				mailSender.send(message);
+			//일정 중복 여부 체크
+			BookMatchingVO db_match = bookService.selectMatch(apply_gatheringDate,apply_num);
+			if(db_match!=null) {
+				mapJson.put("result", "duplicated");
+			}else {
+				//예약 신청 정보 세팅
+				BookMatchingVO bookMatchingVO = new BookMatchingVO();
+				bookMatchingVO.setBook_num(book_num);
+				bookMatchingVO.setApply_num(apply_num);
+				bookMatchingVO.setBook_state(book_state);
+				bookMatchingVO.setApply_gatheringDate(apply_gatheringDate);
 				
-				mapJson.put("result", "apply");
-			}catch(Exception e) {
-				e.printStackTrace();
+				//예약 신청 완료
+				bookService.insertMatch(bookMatchingVO);
+				
+				//이메일 내용 세팅
+				MemberVO member = memberService.selectMember(apply_num);
+				String nickname = member.getMem_nickname();
+				
+				//예약 완료 이메일 양식
+				String setFrom = "229rkawk@gmail.com";
+				String toMail = email;
+				String title = "[ZIBI] 소모임 참여 신청 안내";
+				String content = nickname + "님의 소모임 예약이 완료되었습니다.<br>";
+				content += "변경 사항이 발생할 수 있으니 ZIBI [나의 모임]에서 꼭 확인하세요!";
+				
+				try {
+					MimeMessage message = mailSender.createMimeMessage();
+					MimeMessageHelper helper = new MimeMessageHelper(message, true, "utf-8");
+					helper.setFrom(setFrom);
+					helper.setTo(toMail);
+					helper.setSubject(title);
+					helper.setText(content, true);
+					mailSender.send(message);
+					
+					mapJson.put("result", "apply");
+				}catch(Exception e) {
+					e.printStackTrace();
+				}
 			}
 		}
 		return mapJson;
@@ -114,13 +120,19 @@ public class BookAjaxController {
 	@PostMapping("/book/applyCancel")
 	@ResponseBody
 	public Map<String,String> applyCancel(@RequestParam int book_num,
-										@RequestParam int apply_num){
+										@RequestParam int apply_num,
+										HttpSession session){
 		
 		Map<String,String> mapJson = new HashMap<String,String>();
-		//참여 취소 처리
-		bookService.deleteMatch(book_num, apply_num);
-		mapJson.put("result", "cancel");
 		
+		MemberVO user = (MemberVO)session.getAttribute("user");
+		if(user==null) {
+			mapJson.put("result", "logout");
+		}else {
+			//참여 취소 처리
+			bookService.deleteMatch(book_num, apply_num);
+			mapJson.put("result", "cancel");
+		}
 		return mapJson;
 	}
 
@@ -128,13 +140,19 @@ public class BookAjaxController {
 	@PostMapping("/book/applyApprove")
 	@ResponseBody
 	public Map<String,String> applyApprove(@RequestParam int book_num,
-										@RequestParam int apply_num){
+										@RequestParam int apply_num,
+										HttpSession session){
 		
 		Map<String,String> mapJson = new HashMap<String, String>();
-		//참여 승인 처리
-		bookService.approveMatch(book_num, apply_num);
-		mapJson.put("result", "approve");
 		
+		MemberVO user = (MemberVO)session.getAttribute("user");
+		if(user==null) {
+			mapJson.put("result", "logout");
+		}else {
+			//참여 승인 처리
+			bookService.approveMatch(book_num, apply_num);
+			mapJson.put("result", "approve");
+		}
 		return mapJson;
 	}
 	
@@ -142,25 +160,53 @@ public class BookAjaxController {
 	@PostMapping("/book/applyDeny")
 	@ResponseBody
 	public Map<String,String> denyApprove(@RequestParam int book_num,
-										@RequestParam int apply_num){
+										@RequestParam int apply_num,
+										HttpSession session){
 		
 		Map<String,String> mapJson = new HashMap<String, String>();
-		//참여 거절 처리
-		bookService.denyMatch(book_num, apply_num);
-		mapJson.put("result", "deny");
 		
+		MemberVO user = (MemberVO)session.getAttribute("user");
+		if(user==null) {
+			mapJson.put("result", "logout");
+		}else {
+			//참여 거절 처리
+			bookService.denyMatch(book_num, apply_num);
+			mapJson.put("result", "deny");
+		}
 		return mapJson;
 	}
 	
-	/*-- 모집 완료하기(참여 일괄 거절하기) --*/
-	@PostMapping("/book/complete")
+	/*-- 모집 완료하기(참여 일괄 거절 및 모집 마감하기) --*/
+	@PostMapping("/book/applyClose")
 	@ResponseBody
-	public Map<String,String> applyComplete(@RequestParam int book_num){
+	public Map<String,String> applyClose(@RequestParam int book_num, HttpSession session){
 		Map<String,String> mapJson = new HashMap<String, String>();
-		//모집 완료 처리
-		bookService.updateAllMatch(book_num);
-		mapJson.put("result", "complete");
 		
+		MemberVO user = (MemberVO)session.getAttribute("user");
+		if(user==null) {
+			mapJson.put("result", "logout");
+		}else {
+			//모집 완료 처리
+			bookService.denyAllMatch(book_num);
+			mapJson.put("result", "close");
+		}
+		return mapJson;
+	}
+	
+	/*-- 모임 완료하기 --*/
+	@PostMapping("/book/bookComplete")
+	@ResponseBody
+	public Map<String,String> bookComplete(@RequestParam int book_num, HttpSession session){
+		Map<String,String> mapJson = new HashMap<String, String>();
+		
+		MemberVO user = (MemberVO)session.getAttribute("user");
+		if(user==null) {
+			mapJson.put("result", "logout");
+		}else {
+			//모임 완료 처리
+			bookService.updateOnoff1(book_num);
+			mapJson.put("result", "complete");
+		}
 		return mapJson;
 	}
 }
